requirements.txt
click==8.0.1
google-api-python-client==2.10.0
gpiozero==1.6.2
nltk==3.6.2
oauth2client==4.1.3
PyAudio==0.2.11
python-dotenv==0.18.0
SpeechRecognition==3.8.1

